--- asmt_files/Asmt04_Run1.txt
@@ -1,96 +1,0 @@
------ LINKED BAG 340 C++-----
-
---->>>>> Test 1 --->>>>>
- !add()...     #-END 5-FIVE 4-FOUR 4-FOUR 3-THREE 2-TWO 1-ONE 0-ZERO #-BEGIN
- !Display bag: #-BEGIN 0-ZERO 1-ONE 2-TWO 3-THREE 4-FOUR 4-FOUR 5-FIVE #-END
-  -----------> 9 item(s) total
-
---->>>>> Test 2 --->>>>>
- !removeSecondNode340()...
- !removeSecondNode340()...
- !Display bag: #-BEGIN 2-TWO 3-THREE 4-FOUR 4-FOUR 5-FIVE #-END
-  -----------> 7 item(s) total
-
- !removeSecondNode340()...
- !removeSecondNode340()...
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END
-  -----------> 5 item(s) total
-
---->>>>> Test 3 --->>>>>
- !addEnd340()...
- !addEnd340()...
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR
-  -----------> 7 item(s) total
-
- !addEnd340()...
- !addEnd340()...
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 4 --->>>>>
- !getCurrentSize340Iterative - Iterative...
-  ---> Current size: 9
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 5 --->>>>>
- !getCurrentSize340Recursive() - Recursive...
-  ---> Current size: 9
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 6 --->>>>>
- !getCurrentSize340RecursiveNoHelper() - Recursive...
-  ---> Current size: 9
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 7 --->>>>>
- !getFrequencyOf()...
-  ---> 0-ZERO:  1
-  ---> 1-ONE:   0
-  ---> 2-TWO:   0
-  ---> 4-FOUR:  3
-  ---> 9-NINE:  2
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
- !getFrequencyOf340Recursive() - Recursive...
-  ---> 0-ZERO:  1
-  ---> 1-ONE:   0
-  ---> 2-TWO:   0
-  ---> 4-FOUR:  3
-  ---> 9-NINE:  2
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 8 --->>>>>
- !getFrequencyOf340RecursiveNoHelper() - Recursive...
-  ---> 0-ZERO:  1
-  ---> 1-ONE:   0
-  ---> 2-TWO:   0
-  ---> 4-FOUR:  3
-  ---> 9-NINE:  2
- !Display bag: #-BEGIN 4-FOUR 4-FOUR 5-FIVE #-END 9-NINE 4-FOUR 9-NINE 0-ZERO
-  -----------> 9 item(s) total
-
---->>>>> Test 9 --->>>>>
- !removeRandom340() ---> 4-FOUR
- !removeRandom340() ---> 9-NINE
- !Display bag: 4-FOUR 5-FIVE #-END #-BEGIN 4-FOUR 9-NINE 0-ZERO
-  -----------> 7 item(s) total
-
- !removeRandom340() ---> #-BEGIN
- !removeRandom340() ---> 4-FOUR
- !Display bag: #-END 5-FIVE 4-FOUR 9-NINE 0-ZERO
-  -----------> 5 item(s) total
-
- !removeRandom340() ---> 9-NINE
- !removeRandom340() ---> #-END
- !Display bag: 4-FOUR 5-FIVE 0-ZERO
-  -----------> 3 item(s) total
-
- !removeRandom340() ---> 0-ZERO
- !removeRandom340() ---> 5-FIVE
- !Display bag: 4-FOUR
-  -----------> 1 item(s) total
